refactor: Relocate setCpuFrequency
