+++ code/components/jomjol_mqtt/server_mqtt.cpp
@@ -12,7 +12,7 @@
 #include "interface_mqtt.h"
 #include "time_sntp.h"
 #include "Helper.h"
-#include "system_info.h"
+#include "system.h"
 #include "../../include/defines.h"
