+++ code/main/main.cpp
@@ -10,8 +10,6 @@
 //#include "driver/gpio.h"
 //#include "sdkconfig.h"
 #include "esp_psram.h"
-#include "esp_pm.h"
-
 
 // SD-Card ////////////////////
 //#include "nvs_flash.h"
@@ -40,7 +38,7 @@
     #include "server_mqtt.h"
 #endif //ENABLE_MQTT
 #include "Helper.h"
-#include "system_info.h"
+#include "system.h"
 #include "statusled.h"
 #include "sdcard_check.h"
 
@@ -55,12 +53,6 @@
     #include "soc/soc.h" 
     #include "soc/rtc_cntl_reg.h" 
 #endif
-
-#ifdef DEBUG_ENABLE_SYSINFO
-#if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL( 4, 0, 0 )
-    #include "esp_sys.h"
-#endif
-#endif //DEBUG_ENABLE_SYSINFO
 
 // define `gpio_pad_select_gpip` for newer versions of IDF
 #if (ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(4, 3, 0))
@@ -91,9 +83,8 @@ extern std::string getFwVersion(void);
 extern std::string getHTMLversion(void);
 extern std::string getHTMLcommit(void);
 
-std::vector<std::string> splitString(const std::string& str);
 void migrateConfiguration(void);
-bool setCpuFrequency(void);
+
 
 static const char *TAG = "MAIN";
 
@@ -234,9 +225,9 @@ extern "C" void app_main(void)
     // ********************************************
     setupTime();    // NTP time service: Status of time synchronization will be checked after every cycle (server_tflite.cpp)
 
-    // Set CPU Frequency
+    // Set CPU Frequency (default: 160Mhz)
     // ********************************************
-    setCpuFrequency();
+    setCPUFrequency();
 
     // SD card: Create further mandatory directories (if not already existing)
     // Correct creation of these folders will be checked with function "SDCardCheckFolderFilePresence"
@@ -442,7 +433,8 @@ extern "C" void app_main(void)
 }
 
 
-void migrateConfiguration(void) {
+void migrateConfiguration(void)
+{
     bool migrated = false;
 
     if (!FileExists(CONFIG_FILE)) {
@@ -719,113 +711,4 @@ void migrateConfiguration(void) {
         fclose(pfile);
         LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Config file migrated. Saved backup to " + std::string(CONFIG_FILE_BACKUP));
     }
-}
-
-
-std::vector<std::string> splitString(const std::string& str) {
-    std::vector<std::string> tokens;
- 
-    std::stringstream ss(str);
-    std::string token;
-
-    while (std::getline(ss, token, '\n')) {
-        tokens.push_back(token);
-    }
- 
-    return tokens;
-}
-
-
-/*bool replace_all(std::string& s, std::string const& toReplace, std::string const& replaceWith) {
-    std::string buf;
-    std::size_t pos = 0;
-    std::size_t prevPos;
-    bool found = false;
-
-    // Reserves rough estimate of final size of string.
-    buf.reserve(s.size());
-
-    while (true) {
-        prevPos = pos;
-        pos = s.find(toReplace, pos);
-        if (pos == std::string::npos) {
-            break;
-        }
-        found = true;
-        buf.append(s, prevPos, pos - prevPos);
-        buf += replaceWith;
-        pos += toReplace.size();
-    }
-
-    buf.append(s, prevPos, s.size() - prevPos);
-    s.swap(buf);
-
-    return found;
-}*/
-
-
-bool setCpuFrequency(void) {
-    ConfigFile configFile = ConfigFile(CONFIG_FILE); 
-    std::string cpuFrequency = "160";
-    esp_pm_config_esp32_t  pm_config; 
-
-    if (!configFile.ConfigFileExists()){
-        LogFile.WriteToFile(ESP_LOG_WARN, TAG, "No config file - exit setCpuFrequency");
-        return false;
-    }
-
-    std::vector<std::string> splitted;
-    std::string line = "";
-    bool disabledLine = false;
-    bool eof = false;
-
-
-    /* Load config from config file */
-    while ((!configFile.GetNextParagraph(line, disabledLine, eof) || 
-            (line.compare("[System]") != 0)) && !eof) {}
-    if (eof) {
-        return false;
-    }
-
-    if (disabledLine) {
-        return false;
-    }
-
-    while (configFile.getNextLine(&line, disabledLine, eof) && 
-            !configFile.isNewParagraph(line)) {
-        splitted = ZerlegeZeile(line);
-
-        if (toUpper(splitted[0]) == "CPUFREQUENCY") {
-            cpuFrequency = splitted[1];
-            break;
-        }
-    }
-
-    if (esp_pm_get_configuration(&pm_config) != ESP_OK) {
-        LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Failed to read CPU Frequency");
-        return false;
-    }
-
-    if (cpuFrequency == "160") { // 160 is the default
-        // No change needed
-    }
-    else if (cpuFrequency == "240") {
-        pm_config.max_freq_mhz = 240;
-        pm_config.min_freq_mhz = pm_config.max_freq_mhz;
-        if (esp_pm_configure(&pm_config) != ESP_OK) {
-            LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Failed to set new CPU frequency");
-            return false;
-        }
-    }
-    else {
-        LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Unknown CPU frequency: " + cpuFrequency + 
-                "It must be 160 or 240");
-        return false;
-    }
-
-    if (esp_pm_get_configuration(&pm_config) == ESP_OK) {
-        LogFile.WriteToFile(ESP_LOG_INFO, TAG, std::string("CPU frequency: ") + std::to_string(pm_config.max_freq_mhz) + " MHz");
-    }
-
-    return true;
 }
